COMP: Fix argument types of `std::min` calls in StackTransform

+++ Common/Transforms/itkStackTransform.hxx
@@ -181,8 +181,8 @@ StackTransform<TScalarType, NInputDimensions, NOutputDimensions>::GetSpatialJaco
   /** Get Jacobian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformSpatialJacobianType sjr;
   this->m_SubTransformContainer[subt]->GetSpatialJacobian(ippr, sjr);
@@ -224,8 +224,8 @@ StackTransform<TScalarType, NDimensions, VSplineOrder>::GetJacobianOfSpatialJaco
   /** Get Jacobian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformTypeJacobianOfSpatialJacobianType subjacspjac;
   this->m_SubTransformContainer[subt]->GetJacobianOfSpatialJacobian(ippr, subjacspjac, nonZeroJacobianIndices);
@@ -279,8 +279,8 @@ StackTransform<TScalarType, NDimensions, VSplineOrder>::GetJacobianOfSpatialJaco
   /** Get Jacobian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformSpatialJacobianType               subjac;
   SubTransformTypeJacobianOfSpatialJacobianType subjacspjac;
@@ -344,8 +344,8 @@ StackTransform<TScalarType, NDimensions, VSplineOrder>::GetSpatialHessian(const 
   /** Get Hessian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformSpatialHessianType subhes;
   this->m_SubTransformContainer[subt]->GetSpatialHessian(ippr, subhes);
@@ -391,8 +391,8 @@ StackTransform<TScalarType, NDimensions, VSplineOrder>::GetJacobianOfSpatialHess
   /** Get Hessian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformJacobianOfSpatialHessianType subjaches;
   this->m_SubTransformContainer[subt]->GetJacobianOfSpatialHessian(ippr, subjaches, nzji);
@@ -452,8 +452,8 @@ StackTransform<TScalarType, NDimensions, VSplineOrder>::GetJacobianOfSpatialHess
   /** Get Hessian from right subtransform. */
   const unsigned int subt =
     std::min(m_SubTransformContainer.size() - 1,
-                 static_cast<unsigned int>(
-                   std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
+             static_cast<size_t>(
+               std::max(0, vnl_math::rnd((ipp[ReducedInputSpaceDimension] - m_StackOrigin) / m_StackSpacing))));
 
   SubTransformSpatialHessianType           subhes;
   SubTransformJacobianOfSpatialHessianType subjaches;
